orangepetclinic/backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.39
pydantic==2.10.6
pydantic-settings==2.8.1
psycopg[binary]==3.2.6
python-jose==3.3.0
PyJWT==2.10.1
bcrypt==4.3.0
python-multipart==0.0.20
supabase==2.15.0
razorpay==1.4.2
alembic==1.15.1
email-validator==2.2.0
